requirements.txt
streamlit
groq
cohere
pypdf
langchain
langchain-community
langchain-core
langchain-groq
langchain-cohere
faiss-cpu
python-dotenv